Remove sandbox, it was all unused 68k stuff anyway
