+++ qtBuildSystem/Mu/emuwrapper.cpp
@@ -25,7 +25,6 @@ extern "C"{
 #include "../../src/m68k/m68k.h"
 #include "../../src/pxa260/pxa260.h"
 #include "../../src/armv5te/disasm.h"
-#include "../../src/debug/sandbox.h"
 }
