+++ src/dbvz.c
@@ -11,7 +11,6 @@
 #include "sdCard.h"
 #include "audio/blip_buf.h"
 #include "m68k/m68k.h"
-#include "debug/sandbox.h"
 
 
 #include "dbvzRegisterNames.c.h"
