+++ src/m68k/m68kexternal.h
@@ -6,6 +6,5 @@
 int32_t interruptAcknowledge(int32_t intLevel);
 void emulatorSoftReset(void);
 void flx68000PcLongJump(uint32_t newPc);
-void sandboxOnOpcodeRun(void);
 
 #endif
